extract create project document button
