+++ app/Dashboard/Projects/ProjectDetail/_components/CreateProjectDocument/CreateProjectDocumentButton/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateProjectDocumentButton';
